Id's changed to Long and attributes to private

+++ src/main/java/soe/mdeis/m7/solid/controller/ProveedorController.java
@@ -40,7 +40,7 @@ public class ProveedorController {
    }
 
    @PutMapping("/proveedor/{id}")
-   public ResponseEntity<Proveedor> updateProveedor(@PathVariable int id,
+   public ResponseEntity<Proveedor> updateProveedor(@PathVariable long id,
          @RequestBody Proveedor proveedor) {
       if (StringUtils.isBlank(proveedor.getNombre())) {
          return ResponseEntity.badRequest().body(proveedor);
