+++ src/main/java/soe/mdeis/m7/solid/repository/FabricanteRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import soe.mdeis.m7.solid.model.Fabricante;
 
-public interface FabricanteRepository extends JpaRepository<Fabricante, Integer> {
+
+public interface FabricanteRepository extends JpaRepository<Fabricante, Long> {
+
 
 }
