+++ src/main/java/soe/mdeis/m7/solid/repository/ServicioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import soe.mdeis.m7.solid.model.Servicio;
 
-public interface ServicioRepository extends JpaRepository<Servicio, Integer> {
+public interface ServicioRepository extends JpaRepository<Servicio, Long> {
 
 }
