+++ src/test/java/soe/mdeis/m7/solid/service/AlmacenServiceTest.java
@@ -31,9 +31,9 @@ class AlmacenServiceTest {
    @Test
    @DisplayName("Almacen Service update service")
    void almacenServiceUpdateService() {
-      Mockito.when(service.save(Mockito.any(Almacen.class))).thenReturn(new Almacen(1, "Almacen"));
-      Almacen almacen = new Almacen(0, "Almacen");
-      Almacen savedAlmacen = service.update(1, almacen);
+      Mockito.when(service.save(Mockito.any(Almacen.class))).thenReturn(new Almacen(1l, "Almacen"));
+      Almacen almacen = new Almacen(0l, "Almacen");
+      Almacen savedAlmacen = service.update(1l, almacen);
       assertNotEquals(almacen.getId(), savedAlmacen.getId());
       assertEquals(almacen.getNombre(), savedAlmacen.getNombre());
    }
@@ -41,8 +41,8 @@ class AlmacenServiceTest {
    @Test
    @DisplayName("Almacen Service save Service")
    void almacenServiceSaveService() {
-      Mockito.when(service.save(Mockito.any(Almacen.class))).thenReturn(new Almacen(1, "Almacen"));
-      Almacen almacen = new Almacen(0, "Almacen");
+      Mockito.when(service.save(Mockito.any(Almacen.class))).thenReturn(new Almacen(1l, "Almacen"));
+      Almacen almacen = new Almacen(0l, "Almacen");
       Almacen savedAlmacen = service.save(almacen);
       assertNotEquals(almacen.getId(), savedAlmacen.getId());
       assertEquals(almacen.getNombre(), savedAlmacen.getNombre());
@@ -52,8 +52,8 @@ class AlmacenServiceTest {
    @DisplayName("Almacen Service get All Almacenes")
    void almacenServiceGetAllAlmacenes() {
       Mockito.when(repository.findAll()).thenReturn(Arrays.asList(
-            new Almacen(1, "Almacen 1"),
-            new Almacen(2, "Almacen 2")));
+            new Almacen(1l, "Almacen 1"),
+            new Almacen(2l, "Almacen 2")));
       var list = service.getAll();
       assertTrue(!list.isEmpty());
       assertEquals(2, list.size());
@@ -62,8 +62,8 @@ class AlmacenServiceTest {
    @Test
    @DisplayName("Almacen Service Get Almacen")
    void almacenServiceGetAlmacen() {
-      Mockito.when(repository.findById(1)).thenReturn(Optional.of(new Almacen(1, "Almacen 1")));
-      Mockito.when(repository.findById(2)).thenReturn(Optional.of(new Almacen(2, "Almacen 2")));
+      Mockito.when(repository.findById(1l)).thenReturn(Optional.of(new Almacen(1l, "Almacen 1")));
+      Mockito.when(repository.findById(2l)).thenReturn(Optional.of(new Almacen(2l, "Almacen 2")));
       Almacen a1 = service.get(1).get();
       Almacen a2 = service.get(2).get();
       assertNotEquals(a1, a2);
@@ -74,7 +74,7 @@ class AlmacenServiceTest {
    @Test
    @DisplayName("Almacen ServiceDeleteAlmacen")
    void almacenServiceDeleteAlmacen() {
-      int almacenId = 1;
+      long almacenId = 1l;
       Mockito.doNothing().when(repository).deleteById(almacenId);
       service.delete(almacenId);
       Mockito.verify(repository, Mockito.times(1)).deleteById(almacenId);
